requirements.txt
asgiref==3.4.1
backcall==0.2.0
certifi==2021.5.30
chardet==4.0.0
colorama==0.4.4
decorator==5.0.9
Django==3.2.5
djangorestframework==3.12.4
idna==2.10
ipython==7.25.0
ipython-genutils==0.2.0
jedi==0.18.0
matplotlib-inline==0.1.2
oauthlib==3.1.1
parso==0.8.2
pickleshare==0.7.5
prompt-toolkit==3.0.19
Pygments==2.9.0
PySocks==1.7.1
python-decouple==3.4
pytz==2021.1
requests==2.25.1
requests-oauthlib==1.3.0
six==1.16.0
sqlparse==0.4.1
traitlets==5.0.5
tweepy==3.10.0
typing-extensions==3.10.0.0
urllib3==1.26.6
wcwidth==0.2.5
